+++ src/frontend/apps/impress/src/features/docs/doc-management/api/index.ts
@@ -2,5 +2,6 @@ export * from './useCreateDoc';
 export * from './useDoc';
 export * from './useDocOptions';
 export * from './useDocs';
+export * from './useCreateFavoriteDoc';
 export * from './useUpdateDoc';
 export * from './useUpdateDocLink';